+++ README.md
@@ -13,9 +13,9 @@ Chat program with RabbitMQ
 	````
 ## Работа с проектом
 
-- Для запуска проекта наобходимо выполнить команду `python3 main_2.py`.
-  - Для вывода на экран текста с описанием параметров наобходимо выполнить команду `python3 main_2.py --help`.
-  - Так же поддерживается 2 опциональных аргумента командной строки `python3 main_2.py <ip_address_of_rabbitMQ> <name_of_chat_channel>`.
+- Для запуска проекта наобходимо выполнить команду `python3 main.py`.
+  - Для вывода на экран текста с описанием параметров наобходимо выполнить команду `python3 main.py --help`.
+  - Так же поддерживается 2 опциональных аргумента командной строки `python3 main.py <ip_address_of_rabbitMQ> <name_of_chat_channel>`.
 - Для просмотра подсказки в приложении необходимо ввести команду `$$`.
 - Для отправки сообщения в текущий канал необходимо ввести сообщение и нажать клавишу ввода (сообщение не  может начинаться с управляющего символа `$`).
 - Для смены канала необходимо ввести команду `$! <имя_канала>`.
@@ -38,16 +38,16 @@ Chat program with RabbitMQ
 
 ## Структура проекта
 ```shell
-                main_2.py
+                main.py
                 /       \
                /         \
-        chat_2.py      commander.py
+        chat.py      commander.py
 ```
-В файле `main_2.py` реализованы 
+В файле `main.py` реализованы 
 - обработка аргументов командой строки 
-- связь `commander.py` и `chat_2.py`
+- связь `commander.py` и `chat.py`
 
-В файле `chat_2.py` реализован класс, взаимодействия с сервером rabbitMQ.
+В файле `chat.py` реализован класс, взаимодействия с сервером rabbitMQ.
 - Отправка сообщений
 - Смена канала
 - Менеджмент каналов
